# packages/openapi-generator/tmp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lonestone API",
    "description": "The Lonestone API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "paths": {
...
    "/api/suppliers/register": {
      "post": {
        "operationId": "SuppliersController_register",
        "tags": [
          "Suppliers"
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Data required to register as a supplier",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "shopName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "INPUTS",
                      "TRANSFORMER"
                    ],
                    "title": "SupplierType",
                    "description": "Type of supplier activity"
                  },
                  "latitude": {
                    "type": "number",
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": -180,
                    "maximum": 180
                  },
                  "address": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 255
                  },
                  "neighborhood": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100
                  },
                  "mobileMoneyNumber": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 20
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "CONTACT",
                      "ORDER"
                    ],
                    "title": "SupplierMode",
                    "description": "How buyers interact with this supplier"
                  },
                  "openingHours": {
                    "type": "object",
                    "propertyNames": {
                      "type": "string",
                      "enum": [
                        "monday",
                        "tuesday",
                        "wednesday",
                        "thursday",
                        "friday",
                        "saturday",
                        "sunday"
                      ]
                    },
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "open": {
                          "type": "string",
                          "pattern": "^\\d{2}:\\d{2}$"
                        },
                        "close": {
                          "type": "string",
                          "pattern": "^\\d{2}:\\d{2}$"
                        },
                        "closed": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "open",
                        "close"
                      ],
                      "additionalProperties": false
                    },
                    "required": [
                      "monday",
                      "tuesday",
                      "wednesday",
                      "thursday",
                      "friday",
                      "saturday",
                      "sunday"
                    ],
                    "title": "OpeningHours",
                    "description": "Weekly opening hours with days as keys"
                  }
                },
                "required": [
                  "shopName",
                  "type",
                  "latitude",
                  "longitude",
                  "address",
                  "neighborhood",
                  "mobileMoneyNumber",
                  "mode"
                ],
                "additionalProperties": false,
                "title": "RegisterSupplier",
                "description": "Data required to register as a supplier"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/suppliers/{id}": {
      "get": {
        "operationId": "SuppliersController_findById",
        "tags": [
          "Suppliers"
        ],
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/suppliers/me": {
      "put": {
        "operationId": "SuppliersController_updateMe",
        "tags": [
          "Suppliers"
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Update supplier profile — all fields optional",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "shopName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "INPUTS",
                      "TRANSFORMER"
                    ],
                    "title": "SupplierType",
                    "description": "Type of supplier activity"
                  },
                  "latitude": {
                    "type": "number",
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": -180,
                    "maximum": 180
                  },
                  "address": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 255
                  },
                  "neighborhood": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100
                  },
                  "mobileMoneyNumber": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 20
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "CONTACT",
                      "ORDER"
                    ],
                    "title": "SupplierMode",
                    "description": "How buyers interact with this supplier"
                  },
                  "openingHours": {
                    "type": "object",
                    "propertyNames": {
                      "type": "string",
                      "enum": [
                        "monday",
                        "tuesday",
                        "wednesday",
                        "thursday",
                        "friday",
                        "saturday",
                        "sunday"
                      ]
                    },
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "open": {
                          "type": "string",
                          "pattern": "^\\d{2}:\\d{2}$"
                        },
                        "close": {
                          "type": "string",
                          "pattern": "^\\d{2}:\\d{2}$"
                        },
                        "closed": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "open",
                        "close"
                      ],
                      "additionalProperties": false
                    },
                    "required": [
                      "monday",
                      "tuesday",
                      "wednesday",
                      "thursday",
                      "friday",
                      "saturday",
                      "sunday"
                    ],
                    "title": "OpeningHours",
                    "description": "Weekly opening hours with days as keys"
                  }
                },
                "additionalProperties": false,
                "title": "UpdateSupplier",
                "description": "Update supplier profile — all fields optional"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
...
}